+++ apps/web/modules/onboarding/organization/invite/organization-invite-view.tsx
@@ -4,7 +4,7 @@ import { useRouter } from "next/navigation";
 import posthog from "posthog-js";
 import React from "react";
 
-import { useFlags } from "@calcom/web/modules/feature-flags/hooks";
+import { useFlags } from "@calcom/web/modules/feature-flags/hooks/useFlags";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
 import { Button } from "@calcom/ui/components/button";
